@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61.0", "wheel"]
+requires = ["setuptools>=82.0.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -11,18 +11,16 @@ authors = [
 ]
 readme = "README.md"
 license = {text = "AGPL-3.0"}
-requires-python = ">=3.10"
+requires-python = ">=3.14"
 classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: GNU Affero General Public License v3",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.14",
 ]
 dependencies = [
-    "ansible==3.4.0",
+    "ansible==11.13.0",
     "beautifulsoup4",
     "boto3==1.39.14",
     "Click",
@@ -37,14 +35,15 @@ dependencies = [
     "prometheus-client==0.19.0",
     "PyGithub==1.44.1",
     "pyOpenSSL~=23.2.0",
-    "python-telegram-bot==13.15",
-    "razorpay~=1.2.0",
+    "python-telegram-bot==22.6",
+    "razorpay~=2.0.0",
     "requests<2.32",
     "responses==0.23.1",
     "selenium==4.23.1",
     "semantic-version==2.10.0",
     "sql_metadata==2.10.0",
-    "stripe~=2.56.0",
+    "standard-imghdr==3.13.0", # Needed for ansible, imghdr was removed from standard library
+    "stripe~=3.1.0",
     "tldextract==3.4.4",
     "tomli==2.0.1",
     "tqdm==4.66.3",
@@ -52,7 +51,7 @@ dependencies = [
     "wrapt~=1.15.0",
     "elasticsearch-dsl>=8.0.0,<9.0.0",
     "hcloud==2.2.1",
-    "playwright==1.49.1",
+    "playwright==1.58.0", # Anything >=1.50.0 works
     "prometheus-api-client==0.6.0",
     "pydo==0.24.0",
 ]
