@@ -3,7 +3,7 @@
 
 import datetime
 import json
-from typing import Optional, TypedDict
+from typing import TypedDict
 
 import frappe
 import frappe.utils
@@ -17,14 +17,11 @@ RECORD_FOR: dict[str, str] = {
 	"Error Log": "method",
 }
 
-Counts = TypedDict(
-	"Counts",
-	{
-		"counts": dict[str, int],
-		"date": datetime.date,
-		"total": int,
-	},
-)
+
+class Counts(TypedDict):
+	counts: dict[str, int]
+	date: datetime.date
+	total: int
 
 
 class LogCounter(Document):
@@ -57,7 +54,7 @@ def record_counts():
 def record_for_date(
 	doctype: str = "Error Log",
 	groupby: str = "method",
-	date: Optional[datetime.date] = None,
+	date: datetime.date | None = None,
 ):
 	counts = get_counts(
 		doctype,
@@ -89,7 +86,7 @@ def record_for_date(
 def get_counts(
 	doctype: str = "Error Log",
 	groupby: str = "method",
-	date: Optional[datetime.date] = None,
+	date: datetime.date | None = None,
 ) -> Counts:
 	date_to = date if date else frappe.utils.now_datetime().date()
 	date_from = date_to - datetime.timedelta(days=1)
@@ -99,7 +96,7 @@ def get_counts(
 
 	q = frappe.qb.from_(table)
 	q = q.select(column, Count("*", alias="count"))
-	q = q.where(table.creation[date_from:date_to])
+	q = q.where(table.creation[date_from:date_to])  # type: ignore
 	q = q.groupby(column)
 	q = q.orderby("count", order=Order.desc)
 	r = q.run()
@@ -118,7 +115,7 @@ def get_name(doctype: str, date: datetime.date):
 def top_k(
 	k: int = 5,
 	log_type: str = "Error Log",
-	since: Optional[datetime.date] = None,
+	since: datetime.date | None = None,
 ):
 	if not since:
 		since = frappe.utils.now_datetime().date() - datetime.timedelta(days=30)
